+++ CHANGELOG.md
@@ -1,9 +1,13 @@
 # Changes
 
-## v1.0.0
-* Apr/05/2021
-* Very first release of this toolkit that was tested to support the use of set, list, map and nested tuple attributes in a rule expression.
+## v1.0.2
+* Apr/15/2021
+* Added support for the use of list<TUPLE> in a rule expression.
 
 ## v1.0.1
 * Apr/12/2021
 * Added support for multilevel parenthesis i.e. nested rule expressions.
+
+## v1.0.0
+* Apr/05/2021
+* Very first release of this toolkit that was tested to support the use of set, list, map and nested tuple attributes in a rule expression.